+++ src/bin/add.rs
@@ -5,12 +5,12 @@ use backhand::filesystem::FilesystemHeader;
 use backhand::Squashfs;
 use clap::Parser;
 
-/// Binary to add file to squashfs filesystem
+/// tool to add files to squashfs filesystems
 #[derive(Parser, Debug)]
-#[command(author, version, about, long_about = None)]
+#[command(author, version)]
 struct Args {
     /// Squashfs file
-    input: PathBuf,
+    filesystem: PathBuf,
 
     // Path of file to read, to write into squashfs
     file: PathBuf,
@@ -25,7 +25,7 @@ fn main() {
     let args = Args::parse();
 
     // read of squashfs
-    let file = File::open(args.input).unwrap();
+    let file = File::open(args.filesystem).unwrap();
     let squashfs = Squashfs::from_reader(file).unwrap();
     let mut filesystem = squashfs.into_filesystem().unwrap();